\name{elephantSeal}
\alias{elephantSeal}
\docType{data}
\title{Location of an elephant seal}
\description{
 This is a data frame giving the location of an elephant seal
off the West coast of the US. The data are from Steward and Brillinger.
}
\usage{data(elephantSeal)}
\format{
  A data frame with 73 observations on the following 5 variables.
  \describe{
    \item{\code{day}}{a numeric vector}
    \item{\code{latitude}}{a numeric vector}
    \item{\code{longitude}}{a numeric vector}
    \item{\code{date}}{a POSIXt}
  }
}
\source{
 Brent Stewart and Davidi Brillinger.
ELEPHANT SEAL MOVEMENTS: MODELLING MIGRATION
}
\examples{
data(elephantSeal)
}
\keyword{datasets}
