\name{USCities}
\alias{USCities}
\docType{data}
\title{Locations of many cities in the USA}
\description{
  A collection of city locations in the USA.
}
\usage{data(USCities)}
\format{
  A data frame with 1217 observations on the following 4 variables.
  \describe{
    \item{\code{latitude}}{a numeric vector}
    \item{\code{longitude}}{a numeric vector}
    \item{\code{name}}{a character vector}
    \item{\code{state}}{a character vector}
  }
}
\source{
Unfortunately, I can't quite recall.
}
\examples{
data(USCities)
}
\keyword{datasets}
